change chainId to u64

+++ core/lib/state/src/in_memory.rs
@@ -9,7 +9,7 @@ use zksync_types::{
 use zksync_utils::u256_to_h256;
 
 /// Network ID we use by defailt for in memory storage.
-pub const IN_MEMORY_STORAGE_DEFAULT_NETWORK_ID: u16 = 270;
+pub const IN_MEMORY_STORAGE_DEFAULT_NETWORK_ID: u64 = 270;
 
 /// In-memory storage.
 #[derive(Debug, Default)]
